// web-client/src/models/transactions.ts
export interface BankAccount {
  id: number;
  accountType: string;
  accountHolder: string;
  accountHolderType: string;
  accountName: string;
  bsb: string;
  accountNumber: string;
  currentBalance: string;
  availableBalance: string;
  transactions: Transaction[];
  statementSummary: StatementSummary;
  statementAnalysis: StatementAnalysis[];
  additionalDetails: AdditionalDetails;
}

export interface AdditionalDetails {
  interestRate: string;
  email: string;
  openDate?: string;
  phone: string;
  address: { [key: string]: string };
  bonusInterestRate?: string;
  debitInterestRate?: string;
  interestPaymentFrequency?: string;
  nextInterestPayment?: string;
  overdraftLimit?: string;
  overdraftRemaining?: string;
  overdraftUsed?: string;
  cardholderRelationship?: string;
  creditLimit?: string;
  lastPayment?: string;
  minimumAmountDue?: string;
  minimumAmountDueDate?: string;
  nextPaymentDue?: string;
}

export interface StatementAnalysis {
  analysisCategory: AnalysisCategory;
}

export interface AnalysisCategory {
  name: CategoryEnum;
  analysisPoints: AnalysisPoint[];
  transactionGroups: TransactionGroup[];
}
// ...
export interface Transaction {
  date: string;
  description: Description;
  amount: number;
  balance: string;
  type: TransactionType;
  tags: Tag[];
  transactionHash: string;
  logo: string;
  suburb: string;
}

export enum Description {
  AnUncategorisedDebit = "An uncategorised debit",
  AutomaticDrawingO13439G2GLoansGoodToGoLoans = "AUTOMATIC DRAWING O13439 G2G LOANS GOOD TO GO LOANS",
  BPAYMoneyToTelstra = "BPAY Money to Telstra",
  CtrlinkNewstart = "Ctrlink Newstart",
  DebitExcessInterest = "Debit Excess Interest",
  DirectDebit428198FERRATUM11813938 = "Direct Debit 428198 FERRATUM 11813938",
  DirectDebit843752GymMembership = "Direct Debit 843752 Gym Membership",
  DirectRVSL301500FerratumAustralLoan419913 = "Direct RVSL 301500 Ferratum Austral Loan 419913",
  DirectRVSLToCompany = "Direct RVSL to company",
  EftposWoolworths1111BentleyWaAu = "EFTPOS WOOLWORTHS 1111BENTLEY WA AU",
  InterestCurrentRateIs105 = "Interest - current rate is 1.05%",
  LoanPayment = "Loan payment",
  LoanRepayment = "Loan Repayment",
  MaintenanceFee = "Maintenance Fee",
  Misc = "Misc",
  MoneyFromCrownbet = "Money from Crownbet",
  MoneyMeLoanFunded = "MoneyMe Loan Funded",
  MoneyMePayment = "MoneyMe payment",
  MoneyToCrownbet = "Money to Crownbet",
  PendingTransaction = "Pending Transaction",
  RentPayment = "Rent payment",
  SunshineLoanRepayment = "Sunshine Loan Repayment",
  TransactionDetailsAvailableNextBusinessDay = "Transaction Details Available Next Business Day",
  VisaPendingEFTPOSWOOLWORTHS1111DATETESTWAAU = "Visa Pending: EFTPOS WOOLWORTHS 1111DATETEST WA AU",
  WageFromDeliveryServiceJob = "Wage from delivery service job",
}

export interface Tag {
  thirdParty?: ThirdParty;
  category?: CategoryEnum;
  creditDebit?: CreditDebit;
  lenderType?: LenderType;
  pending?: Pending;
}
// ...
export enum TransactionType {
  DirectDebit = "Direct Debit",
  Empty = "",
  GeneralPayment = "General Payment",
  InboundTransfer = "Inbound Transfer",
  Informational = "Informational",
  MaintenanceFees = "Maintenance Fees",
  OverdrawnFees = "Overdrawn Fees",
  ReversedTransactions = "Reversed Transactions",
  WelfarePayments = "Welfare Payments",
}

export interface StatementSummary {
  openingBalance: string;
  totalCredits: string;
  totalDebits: string;
  closingBalance: string;
  minBalance: string;
  minDayEndBalance: string;
  daysInNegative: number;
  searchPeriodStart: string;
  searchPeriodEnd: string;
  transactionsStartDate: string;
  transactionsEndDate: string;
  creditsCategorised: string;
  creditsUncategorised: string;
  debitsCategorised: string;
  debitsUncategorised: string;
  internalTransferDebits: string;
  internalTransferCredits: string;
  accountMovement: string;
  dayEndBalances: DayEndBalance[];
}

export interface DayEndBalance {
  date: string;
  balance: string;
}